+++ src/cli/update.ts
@@ -1,7 +1,10 @@
 import PackageDeployerConfiguration from "@/PackageDeployerConfiguration";
+import LocalRepositories from "@/repository/LocalRepositories";
+import LocalRepositoryList from "@/repository/LocalRepositoryList";
 import RepositoryList from "@/repository/RepositoryList";
 import { Octokit } from "@octokit/rest";
 import yargs from "yargs";
+import { PackageDeployer } from "..";
 
 /**
  * Update things
@@ -9,7 +12,7 @@ import yargs from "yargs";
 export default async function updateMain(
 	yargs: any,
 	config: PackageDeployerConfiguration,
-	octokit: Octokit
+	octokit: Octokit,
 ) {
 	return yargs.command(
 		"update",
@@ -21,11 +24,27 @@ export default async function updateMain(
 			// Get(locally) or fetch(from github) repository list
 			const repositoryList = await RepositoryList.sync(
 				RepositoryList.defaultConfigurationFile(),
-				octokit
+				octokit,
 			);
 
 			// Save
 			await repositoryList.save();
-		}
+
+			// Read all the repositories at the path
+			const localRepositories = await LocalRepositoryList.fromPath(
+				config.getPackagesPath(),
+			);
+
+			// Pull all the repositories if they are newer on the remote
+			const repositories = new LocalRepositories(
+				config.getPackagesPath(),
+				localRepositories,
+			);
+			await repositories.pullIfNewer();
+
+			// Initialize package deployer and deploy all
+			const pkgDeployer = new PackageDeployer(config);
+			await pkgDeployer.deploy();
+		},
 	);
 }